Refactored Renderer
