--- src/Awin/Tools/CoffeeBreak/Services/CoffeeBreakPreferenceListRenderer/CoffeeBreakPreferenceListRendererInterface.php
@@ -1,10 +1,0 @@
-<?php
-namespace Awin\Tools\CoffeeBreak\Services\CoffeeBreakPreferenceListRenderer;
-
-
-use Awin\Tools\CoffeeBreak\Entity\CoffeeBreakPreference;
-
-interface CoffeeBreakPreferenceListRendererInterface
-{
-    public function render(array $items) : string;
-}
