+++ src/Awin/Tools/CoffeeBreak/Services/Renderer/RendererInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+
+namespace Awin\Tools\CoffeeBreak\Services\Renderer;
+
+
+interface RendererInterface
+{
+        public function renderJson($data) : string;
+        public function renderHtml($data) : string;
+        public function renderXml($data) : string;
+}
